requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
blinker==1.7.0
boto3==1.34.38
botocore==1.34.38
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.109.2
Flask==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
